Abstract some common code into Portal.h

+++ src/main/c++/array/AllocateInJavaGetArray.cpp
@@ -35,8 +35,8 @@
  * Method:    getArraySize
  * Signature: (J)J
  */
-jlong Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArraySize
-  (JNIEnv *, jclass, jlong handle) {
+jlong Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArraySize(
+    JNIEnv *, jclass, jlong handle) {
   const auto& cpp_array = *reinterpret_cast<std::vector<jnibench::FooObject>*>(handle);
   return static_cast<jlong>(cpp_array.size());
 }
@@ -46,8 +46,8 @@ jlong Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArr
  * Method:    getArrays
  * Signature: (J[Ljava/lang/String;[J)V
  */
-void Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArrays
-  (JNIEnv *env, jclass, jlong handle, jobjectArray name_array, jlongArray value_array) {
+void Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArrays(
+    JNIEnv *env, jclass, jlong handle, jobjectArray name_array, jlongArray value_array) {
   jlong* value_array_ptr = env->GetLongArrayElements(value_array, nullptr);
   if (value_array_ptr == nullptr) {
     // exception thrown: OutOfMemoryError
@@ -56,7 +56,7 @@ void Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArra
 
   auto* cpp_array = reinterpret_cast<std::vector<jnibench::FooObject>*>(handle);
   for (jsize i = 0; i < env->GetArrayLength(name_array); i++) {
-    jnibench::FooObject foo_obj = (*cpp_array)[i];
+    jnibench::FooObject foo_obj = (*cpp_array)[static_cast<size_t>(i)];
 
     jstring jname = env->NewStringUTF(foo_obj.GetName().c_str());
     if (jname == nullptr) {
@@ -72,7 +72,7 @@ void Java_com_evolvedbinary_jnibench_common_array_AllocateInJavaGetArray_getArra
       return;
     }
 
-    value_array_ptr[i] = static_cast<jlong>(foo_obj.GetValue());
+    value_array_ptr[static_cast<size_t>(i)] = static_cast<jlong>(foo_obj.GetValue());
   }
 
   env->ReleaseLongArrayElements(value_array, value_array_ptr, JNI_COMMIT);
